requirements.txt
pandas
polars-lts-cpu
psycopg2-binary==2.9.9
connectorx
pyarrow
sqlalchemy